pyproject.toml
[project]
name = "hale-c"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.3",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
]
